Set favorited status

+++ src/main/java/com/af/gamerecs/Entities/UserGame.java
@@ -51,7 +51,7 @@ public class UserGame {
         return game;
     }
 
-    public float getRating() {
+    public int getRating() {
         return rating;
     }
